Fix a visual bug on dashboard in the Darksend panel
the field "Amounts and Rounds" display a Quantis value, from a string
variable using the number of digits specified in settings.
The code was incorrectly assuming this value to be a string containing
all decimal digits, and was trying to remove all theses digits, leading
to erase too many characters.
This commit remove the character suppression.